fixed layout at exit bug with not finding config dir

+++ src/UC_Quantum_Lab/layout.py
@@ -139,7 +139,8 @@ def _layout_at_exit():
         _trigger()
 
 def _exit():
-    for item in os.listdir(_config_dir):
-        # deletes png html or the trigger file from the config dir
-        if item == _trigger_file:
-            os.remove(os.path.join(_config_dir, item))
+    if _config_dir in os.listdir():
+        for item in os.listdir(_config_dir):
+            # deletes png html or the trigger file from the config dir
+            if item == _trigger_file:
+                os.remove(os.path.join(_config_dir, item))
